requirements.txt
PyQt6==6.6.1
Pillow==10.2.0
reportlab==4.1.0
pyinstaller==6.3.0